refactor(server): extract platform boundaries
